Updated workflow, changed dummy prompts for tests (#17)
* Updated workflow for upgrading old components and making new components, changed dummy prompts in tests

This adds the feature for importing old files if user chooses to upgrade an
old component to a web component by implementing a new workflow. It also changes
the tests to add a new dummy variable.

* Added loader and version check

This adds the loader for asynchronous tasks and checks version
before showing the package directory URL to user. It also updates
package.json to save the dependencies.